feat: Add TodoRepository interface
This commit introduces the TodoRepository interface, which extends JpaRepository to handle database operations for the Todo entity. TodoRepository provides methods for performing CRUD (Create, Read, Update, Delete) operations on todo items in the database. This interface will be used to interact with the database and manage todo items within the application.